--- src/main/java/calculator/extractor/DelimitedNumberExtractor.java
@@ -1,7 +1,0 @@
-package calculator.extractor;
-
-import java.util.List;
-
-public interface DelimitedNumberExtractor {
-    List<String> extract(String input);
-}
