+++ src/main/java/calculator/extractor/developed/CustomDelimiterNumberExtractor.java
@@ -1,27 +1,23 @@
 package calculator.extractor.developed;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
 import java.util.Arrays;
 import java.util.List;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 
-public class CustomDelimiterNumberExtractorTest implements DelimitedNumberExtractorTest {
+public class CustomDelimiterNumberExtractor implements DelimitedNumberExtractor {
     private static final String CUSTOM_DELIMITER_PREFIX = "//";
     private static final String CUSTOM_DELIMITER_SUFFIX = "\\n";
 
     @Override
-    public ExtractResultTest extract(String input) {
+    public ExtractResult extract(String input) {
         if (!isCustomDelimiterFormat(input)) {
-            return new ExtractResultTest(false, List.of());
+            return new ExtractResult(false, List.of());
         }
         String customDelimiter = extractCustomDelimiter(input);
         String numberString = extractNumberString(input);
         List<String> numbers = splitNumbers(numberString, customDelimiter);
-        return new ExtractResultTest(true, numbers);
+        return new ExtractResult(true, numbers);
     }
 
     private boolean isCustomDelimiterFormat(String input) {
@@ -57,18 +53,3 @@ public class CustomDelimiterNumberExtractorTest implements DelimitedNumberExtrac
     }
 }
 
-class CustomDelimiterNumberExtractorMethodTest {
-
-    @Test
-    @DisplayName("extract 성공 테스트")
-    void extract_ok() {
-        String input = "//@\\n1@2@3";
-        CustomDelimiterNumberExtractorTest customDelimiterNumberExtractor = new CustomDelimiterNumberExtractorTest();
-        ExtractResultTest extract = customDelimiterNumberExtractor.extract(input);
-        assertEquals(extract.isExtracted(), true);
-        assertEquals(extract.getNumbers().get(0), "1");
-        assertEquals(extract.getNumbers().get(1), "2");
-        assertEquals(extract.getNumbers().get(2), "3");
-
-    }
-}
