+++ src/main/java/calculator/extractor/developed/DelimitedNumberExtractor.java
@@ -1,0 +1,7 @@
+package calculator.extractor.developed;
+
+public interface DelimitedNumberExtractor {
+    ExtractResult extract(String input);
+
+
+}
