--- src/main/java/calculator/extractor/developed/DelimitedNumberExtractorTest.java
@@ -1,7 +1,0 @@
-package calculator.extractor.developed;
-
-public interface DelimitedNumberExtractorTest {
-    ExtractResultTest extract(String input);
-
-
-}
