+++ src/test/java/calculator/calculator/StringCalculatorImplTest.java
@@ -1,8 +1,8 @@
 package calculator.calculator;
 
-import calculator.extractor.developed.CustomDelimiterNumberExtractorTest;
-import calculator.extractor.developed.NumberExtractorChainTest;
-import calculator.extractor.developed.StandardDelimiterNumberExtractorTest;
+import calculator.extractor.developed.CustomDelimiterNumberExtractor;
+import calculator.extractor.developed.NumberExtractorChain;
+import calculator.extractor.developed.StandardDelimiterNumberExtractor;
 import calculator.parser.NumberParser;
 import calculator.parser.StringCalculatorNumberParser;
 import calculator.validationRule.BasicFormatValidationRule;
@@ -38,9 +38,9 @@ class StringCalculatorImplTest {
 //                new CustomDelimitedNumberExtractor()
 //        ));
 
-        NumberExtractorChainTest numberExtractorChain = new NumberExtractorChainTest(
-                List.of(new StandardDelimiterNumberExtractorTest(),
-                        new CustomDelimiterNumberExtractorTest()));
+        NumberExtractorChain numberExtractorChain = new NumberExtractorChain(
+                List.of(new StandardDelimiterNumberExtractor(),
+                        new CustomDelimiterNumberExtractor()));
         NumberParser numberParser = new StringCalculatorNumberParser();
 
         StringCalculatorImpl stringCalculator = new StringCalculatorImpl(inputValidator, numberExtractorChain,
